Update Nixpkgs to avoid corrupted outputs
Due to a Linux 6.5 bug, some of my builds may be corrupted.
Updating Nixpkgs changes glibc-locales, among other things, which
combined with a GC can force a full system rebuild.

Weeks of build cache lost.

The bootloader had to be changed because systemd-boot appears to crash
immediately on boot after the upgrade.