Update nixpkgs
Switch to mainline Linux, since it now has a PCIe driver.

Add Haskell overrides to work around regressions in Nixpkgs.

Update initramfs modules to work with the new Linux build.

Add an all-haskell package to build all of Haskell.
This link farm excludes all packages that are known not to build on x86
or have foreign dependencies that do not support RISC-V.
It also excludes non-default package versions.